+++ src/main/java/ru/dkalchenko/teatime/exception/PersonNotFoundException.java
@@ -1,8 +1,10 @@
 package ru.dkalchenko.teatime.exception;
 
+import java.math.BigInteger;
+
 public class PersonNotFoundException extends RuntimeException {
 
-    public PersonNotFoundException(Long id) {
+    public PersonNotFoundException(BigInteger id) {
         super("Could not find person " + id);
     }
 }
