+++ src/main/java/ru/dkalchenko/teatime/service/PersonService.java
@@ -2,6 +2,7 @@ package ru.dkalchenko.teatime.service;
 
 import ru.dkalchenko.teatime.model.Person;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -9,11 +10,11 @@ public interface PersonService {
 
     List<Person> findAll();
 
-    Optional<Person> findById(Long id);
+    Optional<Person> findById(BigInteger id);
 
     Person save(Person person);
 
     Person update(Person person);
 
-    void deleteById(Long id);
+    void deleteById(BigInteger id);
 }
